update file and delete file in not use
