--- src/components/card/cardwebminar.css
@@ -1,376 +1,0 @@
-.webminar-card {
-    width: 1400px;
-    padding: 10px;
-    margin-left: 50px;
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 1rem;
-    position: relative;
-    top: 100px;
-}
-
-.card-webminar {
-    width: 340px;
-    border-radius: 5px;
-    overflow: hidden;
-    box-shadow: 0px 0px 15px -5px rgba(0, 0, 0, 0.75);
-}
-
-.heading-teks {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 40px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 110px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 190px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks2 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 390px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks2 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 460px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks2 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 540px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks3 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 740px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks3 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 810px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks3 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 890px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks4 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 1100px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks4 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 1170px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks4 {
-    position: absolute;
-    object-fit: cover;
-    top: 170px;
-    left: 1250px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks5 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 40px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks5 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 110px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks5 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 190px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks6 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 390px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks6 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 460px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks6 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 540px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks7 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 740px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks7 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 810px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks7 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 890px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.heading-teks8 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 1100px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headingg-teks8 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 1170px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.headinggg-teks8 {
-    position: absolute;
-    object-fit: cover;
-    top: 570px;
-    left: 1250px;
-    border-radius: 25px;
-    background-color: white;
-    padding-left: 10px;
-    padding-right: 10px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.image-webminar {
-    display: flex;
-    justify-content: center;
-}
-
-.image-webminar img {
-    overflow: hidden;
-    height: 200px;
-}
-
-.card-title-webminar .text-webminar {
-    padding: 20px;
-    font-size: 20px;
-    font-weight: bold;
-    color: #002157;
-}
-
-.card-body-webminar .text-webminar {
-    position: relative;
-    left: 21px;
-    bottom: 10px;
-    color: #848484;
-    font-weight: bold;
-}
-
-.btn-webminar {
-    padding: 10px;
-    display: flex;
-    justify-content: center;
-}
-
-.btn-webminar .button-webminar {
-    background-color: #002157;
-    color: white;
-    font-weight: bold;
-    padding: 10px;
-    padding-left: 100px;
-    padding-right: 100px;
-    border-radius: 5px;
-}
-
-.btn-webminar .button-webminar:hover {
-    background-color: #286cdb;
-}
-
-
