--- src/components/navbar/listabout.json
@@ -1,20 +1,0 @@
-[
-    {
-        "sejarah": "Sejarah konseling satir"
-    },
-    
-    {
-        "tentang": "Tentang kami"
-    },
-
-    {
-        "linimasa": "Linimasa konseling satir"
-    },
-
-    {
-        "pendiri": "Pendiri konseling satir"
-    },
-    {
-        "join": "Cara join layanan"
-    }
-]
