--- src/components/navbar/navbarhome.css
@@ -1,55 +1,0 @@
-.dropdown {
-    position: relative;
-    display: inline-block;
-  }
-  
-  .dropdown-menu {
-    position: absolute;
-    top: 100%;
-    left: 0;
-    z-index: 1000;
-    display: none;
-    width: 100%; /* Menyimpan dropdown item full width */
-    padding: 0.5rem 0;
-    margin: 0.125rem 0 0;
-    font-size: 1rem;
-    color: #212529;
-    text-align: left;
-    background-color: #fff;
-    background-clip: padding-box;
-    border: 1px solid rgba(0, 0, 0, 0.15);
-    border-radius: 0.25rem;
-  }
-  
-  .dropdown-menu a {
-    display: block;
-    width: 100%; /* Menyimpan dropdown item full width */
-    padding: 0.25rem 1.5rem;
-    clear: both;
-    font-weight: 400;
-    text-align: inherit;
-    white-space: nowrap;
-    background-color: transparent;
-    border: 0;
-  }
-  
-  .dropdown-menu a:hover {
-    text-decoration: none;
-    background-color: transparent;
-  }
-  
-  .btn-secondary {
-    background-color: transparent;
-    display: inline-block;
-    color: black;
-    padding: 0.5rem 1rem;
-    text-decoration: none;
-    border-radius: 0.25rem;
-    border: none;
-  }
-  
-  .btn-secondary:hover {
-    color: #212529;
-    background-color: transparent;
-  }
-
