--- src/components/navbar/navbarset.css
@@ -1,15 +1,0 @@
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  height: 100px;
-  min-width: 1400px;
-  margin: 0 20px;
-  position: fixed;
-  z-index: 100;
-  transition: background-color 0.3s ease;
-}
-
-.scrolled {
-  background: linear-gradient(90deg, rgb(139, 43, 43) 0%, rgb(99, 57, 57) 100%);
-}
