+++ src/hooks/useGetUserData.tsx
@@ -2,7 +2,7 @@ import { useEffect, useState } from "react";
 import axios from "axios";
 import { handlerApi } from "../services/handlerApi";
 
-import { getTokenAuth } from "../helper/helper";
+// import { getTokenAuth } from "../helper/helper";
 
 const useGetUserData = () => {
     const [email, setEmail] = useState("");
@@ -10,17 +10,17 @@ const useGetUserData = () => {
     const [isFetchingData, setIsFetchingData] = useState(true);
     const [isLoading, setIsLoading] = useState(true);
 
-    // const fetchData = async () => {
+    //   const fetchData = async () => {
     //     try {
-    //         const res = await handlerApi("/users");
-    //         setEmail(res.email);
-    //         setIsLoggedIn(true);
-    //         setIsFetchingData(false);
+    //       const res = await handlerApi("/users");
+    //       setEmail(res.email);
+    //       setIsLoggedIn(true);
+    //       setIsFetchingData(false);
     //     } catch (error) {
-    //         console.error("Error:", error);
-    //         setIsFetchingData(false);
+    //       console.error("Error:", error);
+    //       setIsFetchingData(false);
     //     }
-    // };
+    //   };
 
     const fetchUserData = async () => {
         try {
@@ -51,18 +51,18 @@ const useGetUserData = () => {
 
     console.log(email);
 
-    // useEffect(() => {
+    //   useEffect(() => {
     //     const accessToken = getTokenAuth();
     //     if (accessToken) {
-    //         fetchUserData();
+    //       fetchUserData(accessToken);
     //     } else {
-    //         setIsFetchingData(false);
+    //       setIsFetchingData(false);
     //     }
     //     console.info("logged ", isLoggedIn);
     //     const loadingTimer = setTimeout(() => {
-    //         setIsLoading(false);
+    //       setIsLoading(false);
     //     }, 2000);
-    // }, [!isLoggedIn]);
+    //   }, [!isLoggedIn]);
 
     return {
         email,
