+++ src/pages/home/home.tsx
@@ -1,11 +1,11 @@
-import Navbar from '../../components/navbar/navbarhome';
+import Navbar from '../../components/navbar/navbar';
 import Banner from '../../components/banner/bnhome';
 import Contents from '../../components/content/content';
 import Teks from '../../components/teks/teks-kata-mereka';
 import Card from '../../components/card/cardhome';
 import Icons from '../../components/iconshome/iconshome';
 import Sosmed from '../../components/sosmed/Founding';
-import Intro from '../../components/intro/introhome';
+import Intro from '../../components/youtube/youtube';
 import Article2 from '../../components/articlehome/article2';
 import Questions from '../../components/questions/questions';
 import Footer from '../../components/footer/footer';
