+++ src/pages/profile/profile.tsx
@@ -1,4 +1,4 @@
-import Navbar from '../../components/navbar/navprofile'
+import Navbar from '../../components/navbar/navbar'
 import Content from '../../components/content/content2'
 import Content2 from '../../components/content/content3'
 import Footer from '../../components/footer/footer'
